Move messages closer to components that use them
