+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/HiddenNavbarItemList/index.tsx
@@ -21,7 +21,7 @@ import Header from './Header';
 // i18n
 import { injectIntl } from 'utils/cl-intl';
 import { InjectedIntlProps } from 'react-intl';
-import messages from '../../messages';
+import messages from './messages';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
