+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/HiddenNavbarItemList/messages.ts
@@ -1,0 +1,9 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  deletePageConfirmationHidden: {
+    id: 'app.modules.navbar.admin.containers.deletePageConfirmationHidden',
+    defaultMessage:
+      'Are you sure you want to delete this page? This cannot be undone.',
+  },
+});
