+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/VisibleNavbarItemList/messages.ts
@@ -1,0 +1,13 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  navigationItems: {
+    id: 'app.modules.navbar.admin.components.NavbarItemList.navigationItems',
+    defaultMessage: 'Pages shown on your navigation bar',
+  },
+  deletePageConfirmationVisible: {
+    id: 'app.modules.navbar.admin.containers.deletePageConfirmation',
+    defaultMessage:
+      'Are you sure you want to delete this page? This cannot be undone. You can also remove the page from the navigation bar if you aren’t ready to delete it yet.',
+  },
+});
