+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/index.tsx
@@ -12,7 +12,7 @@ import HiddenNavbarItemList from './HiddenNavbarItemList';
 
 // i18n
 import { injectIntl, FormattedMessage } from 'utils/cl-intl';
-import messages from '../messages';
+import messages from './messages';
 
 const NavigationSettings = () => (
   <Section>
