+++ front/app/modules/commercial/customizable_navbar/admin/containers/NavigationSettings/messages.ts
@@ -1,0 +1,13 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  pageTitle: {
+    id: 'app.modules.navbar.admin.containers.pageTitle',
+    defaultMessage: 'Navigation bar and platform pages',
+  },
+  pageSubtitle: {
+    id: 'app.modules.navbar.admin.containers.pageSubtitle',
+    defaultMessage:
+      'Your navigation bar can display up to five pages in addition to the Home and All projects pages. You can rename and reorder items and add new pages with your own content.',
+  },
+});
